+++ packages/skeleton-svelte/src/lib/components/Popover/Popover.svelte
@@ -2,17 +2,24 @@
 	import { fade } from 'svelte/transition';
 
 	import * as popover from '@zag-js/popover';
-	import { portal, useMachine, normalizeProps } from '@zag-js/svelte';
+	import { portal, useMachine, normalizeProps, mergeProps } from '@zag-js/svelte';
 	import type { PopoverProps } from './types.js';
 	import { useId } from '$lib/internal/use-id.js';
 
 	let {
 		open = $bindable(false),
 		arrow = false,
+		// Base
+		base = '',
+		classes = '',
 		// Trigger
 		triggerBase = '',
 		triggerBackground = '',
 		triggerClasses = '',
+		// Positioner
+		positionerBase = '',
+		positionerZIndex = '',
+		positionerClasses = '',
 		// Content
 		contentBase = '',
 		contentBackground = '',
@@ -24,6 +31,8 @@
 		// Snippets
 		trigger,
 		content,
+		// Events
+		onclick,
 		// Zag ---
 		...zagProps
 	}: PopoverProps = $props();
@@ -47,15 +56,20 @@
 		}
 	);
 	const api = $derived(popover.connect(snapshot, send, normalizeProps));
+	const triggerProps = $derived(mergeProps(api.getTriggerProps(), { onclick }));
 </script>
 
-<span data-testid="popover">
+<span class="{base} {classes}" data-testid="popover">
 	<!-- Snippet: Trigger -->
-	<button {...api.getTriggerProps()} class="{triggerBase} {triggerBackground} {triggerClasses}">
+	<button {...triggerProps} class="{triggerBase} {triggerBackground} {triggerClasses}">
 		{@render trigger?.()}
 	</button>
 	<!-- Portal -->
-	<div use:portal={{ disabled: !api.portalled }} {...api.getPositionerProps()}>
+	<div
+		use:portal={{ disabled: !api.portalled }}
+		{...api.getPositionerProps()}
+		class="{positionerBase} {positionerZIndex} {positionerClasses}"
+	>
 		<!-- Popover -->
 		{#if api.open}
 			<div {...api.getContentProps()} transition:fade={{ duration: 100 }}>